Localization Cue Correction
Crosstalk is defined as sound from one channel interfering with the sound of
another channel, and in this case it refers to the in-air mixing that occurs
between the left and right speakers over air, as some sound coming from
the left speaker reaches the right ear and vice versa. By calculating a
cancellation waveform and playing it to reduce the level of the
crosstalk, some of the crosstalk can be canceled.

The effect tends to widen the stereo soundstage and should reduce comb
filtering on the phantom center channel that arises from the crosstalk
as the channels create a type of early echo effect with each other
simultaneously.

No effect like this is perfect, as copies of the sound reaches the
user's ears with various time delays by bouncing from room's surfaces.
The time delays don't match. The effect is likely most useful in
strongly treated rooms where reflection paths are already highly
mitigated.