Initial version of Localization Cue Correction
This DSP effect is designed to reduce on-air stereo crosstalk. The
plugin plays inverted, filtered and delayed copies of the sound from the
contralateral channel to produce a cancellation of the crosstalk as the
sound hits the contralateral ear.

This tends to widen the stereo perspective and should also reduce comb
filtering on the phantom center channel due left and right speaker
summing in each ear with different time delays.

Typically, enabling the effect changes the sound of the center channel a
little, and moves panned instruments further laterally apart and around
the listener, increasing the envelopment.

This work was inspired by this similarly named github repository:
https://github.com/MeteorStudioASU/lcc but no code from that repository
was used. I adjusted the parameters to match a more common equilateral
triangle setup and added some filtering to estimate head shadow and a
highpass filter to prevent canceling bass during the processing.

The delay value can be computed by e.g. calculating the sound travel
time difference between left channel and right ear, and right channel
and right ear. It should end up being around 250-400 us, but it varies
by interspeaker distance and the listening distance. Interaural distance
can be assumed to be 21.5 cm for the calculation.

The delay can be measured by playing e.g. pink noise and then looking
for a delay value that seems to result in most attenuation of the noise
in the opposing ear. Similarly, the decay is a free parameter intended
to match the interaural level difference near about 1 kHz. This is
probably best tuned with music. If decay value is too low, the effect is
very subtle, and if it's too high, the decay exceeds cancellation and
begins to create phantom copies of the sound in the wrong side. The
value is the gain of the filtered delay line that is applied on every
loop of the noise through the line, but the filter sum already has some
negative gain, and so the parameter doesn't have an easily defined
meaning.

I've currently soft-disabled the center channel effect by setting its
gain to -99 dB by default.  Experimentally, gain values between -20 to
-3 dB are probably reasonable and reduce the strength of the stereo
widening somewhat. It is an idea from the original project, but I am
considering its removal.