Switch to standard Kemar head
A well-characterized industry standard model is probably the most
defensible choice for now. The data is quite different, but reasonable
1.4m far field 30 degree angle picture was produced and digitized and
ILD modeled with some filters.

Effect is stronger now because ILD is estimated to be less. For now,
I must experiment with this model and see if it proves to be acceptable.