Use official formula for phase delay and optimize
I designed a crude parameter hunter that scans the allpass params
and optimizes either phase angle or phase delay, and either rms
or max error.

I decided to minimize the maximum time error. This is nearly the
same thing as the phase delay except it weighs low frequencies more.
Phase delay is bounded to about 46 us, which is in order of 1.5 cm.
Phase angle maximum error is still about 20 degrees only.